requirements.txt
logging
requests
azure-functions
azure-core
azure-identity
python-telegram-bot
python-telegram-bot[job-queue]
beautifulsoup4
datetime